Finished 15/15 CFG/DFG tests

+++ muon-app/src/test/java/muon/app/common/local/LocalFileSystemTest.java
@@ -15,19 +15,22 @@ public class LocalFileSystemTest {
         //Create Test Directory and File
         new File("srctestdirectory").mkdirs();
         new File("srctestdirectory/filename.txt").createNewFile();
+        new File("emptyDir").mkdirs();
 
     }
 
     public void cleanTestDirs(){
-        String dirs[] = {"srctestdirectory"};
+        String dirs[] = {"srctestdirectory","emptyDir"};
         for(int i=0; i<dirs.length;i++){
-            File index = new File(dirs[i]);
-            String[]entries = index.list();
-            for(String s: entries){
-                File currentFile = new File(index.getPath(),s);
-                currentFile.delete();
+            if(checkFileExists(dirs[i])) {
+                File index = new File(dirs[i]);
+                String[] entries = index.list();
+                for (String s : entries) {
+                    File currentFile = new File(index.getPath(), s);
+                    currentFile.delete();
+                }
+                index.delete();
             }
-            index.delete();
         }
     }
 
@@ -42,7 +45,25 @@ public class LocalFileSystemTest {
 
 
     @Test
-    public void list() throws Exception{
+    public void list1() throws Exception{
+        testDirSetup();
+        LocalFileSystem lfs = new LocalFileSystem();
+        List<FileInfo> flist =  lfs.list("");
+        assertTrue(flist.size()>0);
+        cleanTestDirs();
+    }
+
+    @Test
+    public void list2() throws Exception{
+        testDirSetup();
+        LocalFileSystem lfs = new LocalFileSystem();
+        List<FileInfo> flist =  lfs.list("srctestdirectory/");
+        assertEquals(flist.get(0).getName(),"filename.txt");
+        cleanTestDirs();
+    }
+
+    @Test
+    public void list3() throws Exception{
         testDirSetup();
         LocalFileSystem lfs = new LocalFileSystem();
         List<FileInfo> flist =  lfs.list("srctestdirectory");
@@ -51,7 +72,7 @@ public class LocalFileSystemTest {
     }
 
     @Test
-    public void delete() throws Exception{
+    public void testDelete1() throws Exception{
         testDirSetup();
         LocalFileSystem lfs = new LocalFileSystem();
         FileInfo fi = new FileInfo("filename.txt", "srctestdirectory/filename.txt", 0, FileType.File,
@@ -61,4 +82,29 @@ public class LocalFileSystemTest {
         assertFalse(checkFileExists("srctestdirectory/filename.txt"));
         cleanTestDirs();
     }
+
+    @Test
+    public void testDelete2() throws Exception{
+        testDirSetup();
+        LocalFileSystem lfs = new LocalFileSystem();
+        FileInfo fi = new FileInfo("emptyDir", "emptyDir", 0, FileType.Directory,
+                2, 0, "","", 1, "", false);
+
+        lfs.delete(fi);
+        assertFalse(checkFileExists("emptyDir"));
+        cleanTestDirs();
+    }
+
+    @Test
+    public void testDelete3() throws Exception{
+        testDirSetup();
+        LocalFileSystem lfs = new LocalFileSystem();
+        FileInfo fi = new FileInfo("srctestdirectory", "srctestdirectory", 0, FileType.Directory,
+                2, 0, "","", 1, "", false);
+
+        lfs.delete(fi);
+        assertFalse(checkFileExists("srctestdirectory"));
+        cleanTestDirs();
+    }
+
 }
